Lab1_Funnel/requirements.txt
psycopg2-binary==2.9.3
pandas==1.3.5